Add explicit casts between scalar types
* Some functions accept only string, cast ints and floats to string
* After preg_matches or explode() casts numbers to int to do maths
* Cast unix timestamps to int to do maths
* Cast return values from timestamp format function to int
* Cast bitwise operator to bool when needed as bool

* php internal functions like floor/round/ceil documented to return
  float, most cases the result is used as int, added casts

Found by phan strict checks

+++ includes/media/ThumbnailImage.php
@@ -73,8 +73,8 @@ class ThumbnailImage extends MediaTransformOutput {
 		# These should be integers when they get here.
 		# If not, there's a bug somewhere.  But let's at
 		# least produce valid HTML code regardless.
-		$this->width = round( $actualParams['width'] );
-		$this->height = round( $actualParams['height'] );
+		$this->width = (int)round( $actualParams['width'] );
+		$this->height = (int)round( $actualParams['height'] );
 
 		$this->page = $actualParams['page'];
 		$this->lang = $actualParams['lang'];
